prj-str, CORS, sqlite, '/log', req.txt
