+++ Deploy.md
@@ -1,0 +1,8 @@
+### Stetps to deploy:
+1. visit  **Render** (Best Option for Flask)
+2. Push your Flask project to GitHub. (DONE)
+3. Create an account on [Render](https://render.com/)
+4. Click "**New Web Service**", connect your GitHub repo.
+    - set build command as `pip install -r requirements.txt`
+    - set start command as `flask app`
+5. Click **Deploy**!
