--- index.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Fake News Detector</title>
-    <!-- Correct link to CSS in the static/css folder -->
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
-</head>
-<body>
-    <div class="container">
-        <h1>📰 Fake News Detector</h1>
-        <p class="subtext">Analyze articles and determine their authenticity.</p>
-
-        <form id="newsForm" method="POST" action="/check">
-            <div class="input-container">
-                <textarea name="article" placeholder="Paste the news article here..." required></textarea>
-            </div>
-            <button type="submit">🔍 Analyze</button>
-        </form>
-
-        <!-- Loading animation -->
-        <div class="loading" id="loading">
-            <div></div><div></div><div></div>
-        </div>
-
-        <!-- Display result -->
-        {% if result %}
-        <div id="result" class="result show {{ result|lower }}">
-            <h3 class="verdict">This article appears to be: <span class="highlight">{{ result }}</span></h3>
-            <p class="confidence">Confidence: {{ confidence }}</p>
-        </div>
-        {% endif %}
-    </div>
-
-    <!-- Correct link to JS in the static/js folder -->
-    <script src="{{ url_for('static', filename='js/script.js') }}"></script>
-</body>
-</html>
